Conflicts:
	config/routes.rb
	db/schema.rb

+++ app/controllers/upload_controller.rb
@@ -1,4 +1,5 @@
 class UploadController < ApplicationController
+  skip_filter :verify_authenticity_token, :destroy
   def index
      
    @files = DataFile.all
@@ -14,10 +15,9 @@ def create
   
    @post = DataFile.save(params[:file])
    if @post
-    #@filename = params[:file].original_filename
-    @file = DataFile.last
+     @file = DataFile.last
    end
-   render "fileSaved"      # data:remote funktioniert irgendwie nur im Wiki !?
+   render "fileSaved"      
 end
 
 def show
@@ -25,19 +25,23 @@ def show
   
 end
 def destroy
-  @id = params[:id].to_i
-  @file = DataFile.find(params[:id])
-  if File.exist?(Rails.root.join('public', 'uploads', @file.getStorableName))
-    File.delete(Rails.root.join('public', 'uploads', @file.getStorableName))
-  end
-  @file.destroy
-     #TODO: JS Respond
-  #render "index"
-  respond_to do |format|
+  if session[:admin].nil?
+    respond_to do |format|
+      format.js {render js: "alert('Nur Amdins duerfen Dateien loeschen!');"}
+      format.html {render text: "Nur Administratoren duerfen Dateien loeschen!"}
+    end  
+  else  
+    @id = params[:id].to_i
+    @file = DataFile.find(params[:id])
+    if File.exist?(Rails.root.join('public', 'uploads', @file.getStorableName))
+      File.delete(Rails.root.join('public', 'uploads', @file.getStorableName))
+    end
+    @file.destroy
+    respond_to do |format|
       format.js {render "destroy.js.erb"}
       format.html {index}
     end
+  end
 end
-  
   
 end
